requirements.txt
Django==5.1.*
gunicorn==22.0.*
psycopg[binary]==3.2.*
redis==5.0.*
Pillow==10.4.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
djangorestframework-simplejwt==5.3.*